feat: Add BDS docs and refactor command structure
This commit introduces several major changes and improvements to the addon.

**1. Bedrock Dedicated Server (BDS) Support:**
- Added `world_behavior_packs.json` and `world_resource_packs.json` to the root directory to facilitate easier installation on BDS.
- Updated `Docs/F.A.Q.md` with a new section detailing the installation process for BDS, including instructions on how to place the pack folders, use the new JSON files, and enable the required Beta APIs.

**2. Command Refactoring:**
- To improve organization and reduce file clutter, many related command files have been merged into single files.
- The following groups of commands were combined:
  - `home`, `homes`, `sethome`, `delhome`
  - `tpa`, `tpahere`, `tpaccept`, `tpadeny`, `tpacancel`
  - `report`, `reports`, `clearreports`
  - `rbounty`, `bounty`, `listbounty`
  - `pay`, `payconfirm`
  - `spawn`, `setspawn`
  - `ban`, `offlineban`
- The command loader (`index.js`) has been updated to reflect these changes, and all redundant files have been deleted.

**3. Linting:**
- Fixed a minor linting issue in `dataManager.js` where the `world` import was unused.
