--- AddonExeBP/scripts/modules/commands/clearreports.js
@@ -1,15 +1,0 @@
-import { commandManager } from './commandManager.js';
-import { clearAllReports } from '../../core/reportManager.js';
-
-commandManager.register({
-    name: 'clearreports',
-    description: 'Clears all active reports.',
-    category: 'Moderation',
-    permissionLevel: 1, // Admin and above
-    allowConsole: true,
-    parameters: [],
-    execute: (player, args) => {
-        clearAllReports();
-        player.sendMessage('§aAll reports have been cleared.');
-    }
-});
