--- AddonExeBP/scripts/modules/commands/delhome.js
@@ -1,26 +1,0 @@
-import { commandManager } from './commandManager.js';
-import * as homesManager from '../../core/homesManager.js';
-import { getConfig } from '../../core/configManager.js';
-
-commandManager.register({
-    name: 'delhome',
-    aliases: ['remhome', 'deletehome', 'rmhome', '-home'],
-    description: 'Deletes one of your set homes.',
-    category: 'Home System',
-    permissionLevel: 1024, // Everyone
-    parameters: [
-        { name: 'homeName', type: 'string', description: 'The name of the home to delete.' }
-    ],
-    execute: (player, args) => {
-        const config = getConfig();
-        if (!config.homes.enabled) {
-            player.sendMessage('§cThe homes system is currently disabled.');
-            return;
-        }
-
-        const { homeName } = args;
-
-        const result = homesManager.deleteHome(player, homeName);
-        player.sendMessage(result.success ? `§a${result.message}` : `§c${result.message}`);
-    }
-});
