--- AddonExeBP/scripts/modules/commands/tpacancel.js
@@ -1,20 +1,0 @@
-import { commandManager } from './commandManager.js';
-import * as tpaManager from '../../core/tpaManager.js';
-import { getConfig } from '../../core/configManager.js';
-
-commandManager.register({
-    name: 'tpacancel',
-    description: 'Cancels your outgoing TPA request.',
-    category: 'TPA System',
-    permissionLevel: 1024, // Everyone
-    parameters: [],
-    execute: (player, args) => {
-        const config = getConfig();
-        if (!config.tpa.enabled) {
-            player.sendMessage('§cThe TPA system is currently disabled.');
-            return;
-        }
-
-        tpaManager.cancelRequest(player);
-    }
-});
