--- AddonExeBP/scripts/modules/commands/tpadeny.js
@@ -1,21 +1,0 @@
-import { commandManager } from './commandManager.js';
-import * as tpaManager from '../../core/tpaManager.js';
-import { getConfig } from '../../core/configManager.js';
-
-commandManager.register({
-    name: 'tpadeny',
-    aliases: ['tpno', 'tpdeny'],
-    description: 'Denies an incoming TPA request.',
-    category: 'TPA System',
-    permissionLevel: 1024, // Everyone
-    parameters: [],
-    execute: (player, args) => {
-        const config = getConfig();
-        if (!config.tpa.enabled) {
-            player.sendMessage('§cThe TPA system is currently disabled.');
-            return;
-        }
-
-        tpaManager.denyRequest(player);
-    }
-});
