+++ world_behavior_packs.json
@@ -1,0 +1,6 @@
+[
+    {
+        "pack_id": "1564bfa3-6796-4b67-826a-03ca1ac2f3c1",
+        "version": [0, 0, 0]
+    }
+]
