+++ world_resource_packs.json
@@ -1,0 +1,6 @@
+[
+    {
+        "pack_id": "9dca0c93-f6ec-4701-a74c-a774600640a4",
+        "version": [0, 0, 0]
+    }
+]
